% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%notin\%}
\alias{\%notin\%}
\title{x %notin% y == !x %in% y 
Note that, unlike %i% and %u%, but like %in%, %notin% is not commutative.}
\usage{
x \%notin\% table
}
\description{
x %notin% y == !x %in% y 
Note that, unlike %i% and %u%, but like %in%, %notin% is not commutative.
}
