% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect.R
\name{intersect}
\alias{intersect}
\title{A cheap trick, but not enough people know it!}
\usage{
intersect(...)
}
\arguments{
\item{...}{As many intersect-able elements as you wish.}
}
\value{
Their intersection (or an error, if they can't be intersected)
}
\description{
A cheap trick, but not enough people know it!
}
\seealso{
base::funprog

base::intersect
}
