% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{Rscripts-operators}
\alias{Rscripts-operators}
\alias{\%notin\%}
\alias{\%d\%}
\alias{\%i\%}
\alias{\%u\%}
\title{Additional operators which I find useful (especially for chaining)}
\usage{
x \%notin\% table

x \%d\% y

x \%i\% y

x \%u\% y
}
\arguments{
\item{x}{the left hand side of x (grapes) y}

\item{y}{the right hand side of x (grapes) y}
}
\value{
whatever the result of x (grapes) y might be
}
\description{
x %i% y == intersect(x, y)
x %u% y == union(x, y)
}
\details{
x %u% y == setdiff(x, y)
Note that unlike %i% and %u%, %d% is not commutative.

x %notin% y == !x %in% y 
Note that unlike %i% and %u% (but like %in%), %notin% is not commutative.
}
