% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.packages.R
\name{write.packages}
\alias{write.packages}
\title{write a table of all current packages to a CSV file}
\usage{
write.packages(filename = NULL, path = ".")
}
\arguments{
\item{filename}{filename for the results (default is R.major.minor.pkgs.csv)}

\item{path}{path for the results (default is .)}
}
\value{
status of writing the file
}
\description{
write a table of all current packages to a CSV file
}
